@@ -69,11 +69,9 @@ class midcom_services_indexer implements EventSubscriberInterface
      * Support of batch-indexing using an Array of documents instead of a single
      * document is possible (and strongly advised for performance reasons).
      *
-     * @param mixed $documents One or more documents to be indexed, so this is either a
-     *           midcom_services_indexer_document or an Array of these objects.
      * @return boolean Indicating success.
      */
-    public function index($documents) : bool
+    public function index(midcom_services_indexer_document|array $documents) : bool
     {
         if ($this->_disabled) {
             return true;
